:sparkles: add getting page count from a local input source
