+++ .github/workflows/_static-analysis.yml
@@ -13,7 +13,7 @@ jobs:
           submodules: recursive
 
       - name: Set up JDK
-        uses: actions/setup-java@v4
+        uses: actions/setup-java@v5
         with:
           java-version: "11"
           distribution: "temurin"
