+++ src/test/java/com/mindee/extraction/InvoiceSplitterAutoExtractionIT.java
@@ -42,9 +42,7 @@ public class InvoiceSplitterAutoExtractionIT {
   protected PredictResponse<InvoiceV4> getInvoicePrediction(LocalInputSource invoicePDF) throws
     IOException, MindeeException {
     return client.parse(InvoiceV4.class, invoicePDF);
-
   }
-
 
   protected String prepareInvoiceReturn(String rstFilePath, Document<InvoiceV4> invoicePrediction)
     throws IOException {
@@ -60,7 +58,7 @@ public class InvoiceSplitterAutoExtractionIT {
   }
 
   @Test
-  public void givenAPDF_shouldExtractInvoicesStrict() throws IOException, InterruptedException {
+  public void givenAPDF_shouldExtractInvoices() throws IOException, InterruptedException {
     Document<InvoiceSplitterV1> document = getInvoiceSplitterPrediction();
     InvoiceSplitterV1 inference = document.getInference();
 
@@ -71,29 +69,37 @@ public class InvoiceSplitterAutoExtractionIT {
     Assertions.assertEquals(2, extractedPDFsStrict.size());
     Assertions.assertEquals("default_sample_001-001.pdf", extractedPDFsStrict.get(0).getFilename());
     Assertions.assertEquals("default_sample_002-002.pdf", extractedPDFsStrict.get(1).getFilename());
-    PredictResponse<InvoiceV4> invoice0 =
-      getInvoicePrediction(extractedPDFsStrict.get(0).asInputSource());
 
+    PredictResponse<InvoiceV4> invoice0 = getInvoicePrediction(
+        extractedPDFsStrict.get(0).asInputSource()
+    );
     String testStringRSTInvoice0 = prepareInvoiceReturn(
       "src/test/resources/products/invoices/response_v4/summary_full_invoice_p1.rst",
-      invoice0.getDocument());
-    Assertions.assertEquals(testStringRSTInvoice0, String.join(String.format("%n"),
-      invoice0.getDocument().toString().split(System.lineSeparator())));
+      invoice0.getDocument()
+    );
+    double invoice0Ratio = levenshteinRatio(
+        testStringRSTInvoice0,
+        String.join(
+            String.format("%n"),
+            invoice0.getDocument().toString().split(System.lineSeparator())
+        )
+    );
+    Assertions.assertTrue(invoice0Ratio > 0.90);
 
-    PredictResponse<InvoiceV4> invoice1 =
-      getInvoicePrediction(extractedPDFsStrict.get(1).asInputSource());
-
+    PredictResponse<InvoiceV4> invoice1 = getInvoicePrediction(
+        extractedPDFsStrict.get(1).asInputSource()
+    );
     String testStringRSTInvoice1 = prepareInvoiceReturn(
       "src/test/resources/products/invoices/response_v4/summary_full_invoice_p2.rst",
-      invoice1.getDocument());
-    Assertions.assertTrue(
-      levenshteinRatio(
+      invoice1.getDocument()
+    );
+    double invoice1Ratio = levenshteinRatio(
         testStringRSTInvoice1,
-        String.join(String.format("%n"),
-          invoice1.getDocument().toString().split(System.lineSeparator())
+        String.join(
+            String.format("%n"),
+            invoice1.getDocument().toString().split(System.lineSeparator())
         )
-      ) > 0.97);
-
-
+    );
+    Assertions.assertTrue(invoice1Ratio > 0.90);
   }
 }
